requirements.txt
requests==2.31.0
opencv-python==4.8.1.78
numpy==1.24.4
python-dotenv==1.0.0